Cleanup platform headers in all backends (#2499)
* Cleanup platform headers in all backends

(cherry picked from commit 4f50e5cd9f8fba8b9f202fc4e05ca7c5a189490a)
