+++ src/backend/common/jit/Node.hpp
@@ -8,6 +8,7 @@
  ********************************************************/
 
 #pragma once
+#include <af/defines.h>
 #include <optypes.hpp>
 #include <platform.hpp>
