+++ src/backend/cuda/platform.hpp
@@ -9,19 +9,22 @@
 
 #pragma once
 
-#include <GraphicsResourceManager.hpp>
-#include <cublas.hpp>
 #include <cuda.h>
 #include <cuda_runtime.h>
-#include <cufft.hpp>
-#include <cusolverDn.hpp>
-#include <cusparse.hpp>
-#include <memory.hpp>
-
-#include <memory>
 #include <string>
-#include <utility>
-#include <vector>
+
+/* Forward declarations of Opaque structure holding
+ * the following library contexts
+ *  * cuBLAS
+ *  * cuSparse
+ *  * cuSolver
+ */
+struct cublasContext;
+typedef struct cublasContext* BlasHandle;
+struct cusparseContext;
+typedef struct cusparseContext* SparseHandle;
+struct cusolverDnContext;
+typedef struct cusolverDnContext* SolveHandle;
 
 namespace spdlog {
 class logger;
@@ -32,6 +35,12 @@ class ForgeManager;
 }
 
 namespace cuda {
+
+class GraphicsResourceManager;
+class MemoryManager;
+class MemoryManagerPinned;
+class PlanCache;
+
 int getBackend();
 
 std::string getDeviceInfo();
@@ -39,6 +48,7 @@ std::string getDeviceInfo(int device);
 
 std::string getPlatformInfo();
 
+std::string int_version_to_string(int version);
 std::string getDriverVersion();
 
 std::string getCUDARuntimeVersion();
@@ -74,16 +84,8 @@ cudaDeviceProp getDeviceProp(int device);
 
 std::pair<int, int> getComputeCapability(const int device);
 
-struct cudaDevice_t {
-    cudaDeviceProp prop;
-    size_t flops;
-    int nativeId;
-};
-
 bool& evalFlag();
 
-///////////////////////// BEGIN Sub-Managers ///////////////////
-//
 MemoryManager& memoryManager();
 
 MemoryManagerPinned& pinnedMemoryManager();
@@ -99,84 +101,5 @@ BlasHandle blasHandle();
 SolveHandle solverDnHandle();
 
 SparseHandle sparseHandle();
-//
-///////////////////////// END Sub-Managers /////////////////////
 
-class DeviceManager {
-   public:
-    static const size_t MAX_DEVICES = 16;
-
-    static bool checkGraphicsInteropCapability();
-
-    static DeviceManager& getInstance();
-    ~DeviceManager();
-
-    spdlog::logger* getLogger();
-
-    friend MemoryManager& memoryManager();
-
-    friend MemoryManagerPinned& pinnedMemoryManager();
-
-    friend graphics::ForgeManager& forgeManager();
-
-    friend GraphicsResourceManager& interopManager();
-
-    friend std::string getDeviceInfo(int device);
-
-    friend std::string getPlatformInfo();
-
-    friend std::string getDriverVersion();
-
-    friend std::string getCUDARuntimeVersion();
-
-    friend std::string getDeviceInfo();
-
-    friend int getDeviceCount();
-
-    friend int getDeviceNativeId(int device);
-
-    friend int getDeviceIdFromNativeId(int nativeId);
-
-    friend cudaStream_t getStream(int device);
-
-    friend int setDevice(int device);
-
-    friend cudaDeviceProp getDeviceProp(int device);
-
-    friend std::pair<int, int> getComputeCapability(const int device);
-
-   private:
-    DeviceManager();
-
-    // Following two declarations are required to
-    // avoid copying accidental copy/assignment
-    // of instance returned by getInstance to other
-    // variables
-    DeviceManager(DeviceManager const&);
-    void operator=(DeviceManager const&);
-
-    // Attributes
-    enum sort_mode { flops = 0, memory = 1, compute = 2, none = 3 };
-
-    void checkCudaVsDriverVersion();
-    void sortDevices(sort_mode mode = flops);
-
-    int setActiveDevice(int device, int native = -1);
-
-    std::shared_ptr<spdlog::logger> logger;
-
-    std::vector<cudaDevice_t> cuDevices;
-    std::vector<std::pair<int, int>> devJitComputes;
-
-    int nDevices;
-    cudaStream_t streams[MAX_DEVICES];
-
-    std::unique_ptr<graphics::ForgeManager> fgMngr;
-
-    std::unique_ptr<MemoryManager> memManager;
-
-    std::unique_ptr<MemoryManagerPinned> pinnedMemManager;
-
-    std::unique_ptr<GraphicsResourceManager> gfxManagers[MAX_DEVICES];
-};
 }  // namespace cuda
