+++ src/backend/cuda/sparse_arith.cu
@@ -18,6 +18,7 @@
 #include <common/err_common.hpp>
 #include <complex.hpp>
 #include <copy.hpp>
+#include <cusparse.hpp>
 #include <lookup.hpp>
 #include <math.hpp>
 #include <platform.hpp>
@@ -26,11 +27,11 @@
 namespace cuda {
 
 using namespace common;
-using namespace std;
+using std::numeric_limits;
 
 template<typename T>
 T getInf() {
-    return scalar<T>(std::numeric_limits<T>::infinity());
+    return scalar<T>(numeric_limits<T>::infinity());
 }
 
 template<>
