+++ src/backend/cuda/sparse_blas.cpp
@@ -8,6 +8,7 @@
  ********************************************************/
 
 #include <cuda_runtime.h>
+#include <cusparse.hpp>
 #include <platform.hpp>
 #include <sparse_blas.hpp>
 
@@ -18,8 +19,6 @@
 #include <string>
 
 namespace cuda {
-
-using namespace std;
 
 cusparseOperation_t toCusparseTranspose(af_mat_prop opt) {
     cusparseOperation_t out = CUSPARSE_OPERATION_NON_TRANSPOSE;
