+++ src/backend/cuda/surface.cpp
@@ -10,6 +10,7 @@
 #include <Array.hpp>
 #include <GraphicsResourceManager.hpp>
 #include <debug_cuda.hpp>
+#include <device_manager.hpp>
 #include <err_cuda.hpp>
 #include <surface.hpp>
