+++ src/backend/cuda/vector_field.cpp
@@ -10,6 +10,7 @@
 #include <Array.hpp>
 #include <GraphicsResourceManager.hpp>
 #include <debug_cuda.hpp>
+#include <device_manager.hpp>
 #include <err_cuda.hpp>
 #include <vector_field.hpp>
