+++ src/backend/opencl/platform.hpp
@@ -8,6 +8,7 @@
  ********************************************************/
 
 #pragma once
+
 #define CL_HPP_ENABLE_EXCEPTIONS
 #define CL_HPP_MINIMUM_OPENCL_VERSION 120
 #define CL_HPP_TARGET_OPENCL_VERSION 120
@@ -20,13 +21,8 @@
 #include <CL/cl2.hpp>
 #pragma GCC diagnostic pop
 
-#include <memory>
+#include <af/opencl.h>
 #include <string>
-#include <vector>
-
-#include <GraphicsResourceManager.hpp>
-#include <cache.hpp>
-#include <memory.hpp>
 
 namespace boost {
 template<typename T>
@@ -41,15 +37,19 @@ namespace graphics {
 class ForgeManager;
 }
 
-// Forward declaration from clFFT.h
-struct clfftSetupData_;
-typedef clfftSetupData_ clfftSetupData;
-
 namespace opencl {
-// Forward declaration from clfft.hpp
-class PlanCache;
 
-struct kc_entry_t;
+// Forward declarations
+class GraphicsResourceManager;
+struct kc_entry_t;  // kernel cache entry
+class MemoryManager;
+class MemoryManagerPinned;
+class PlanCache;  // clfft
+
+static inline bool verify_present(std::string pname, const char* ref) {
+    return pname.find(ref) != std::string::npos;
+}
+
 int getBackend();
 
 std::string getDeviceInfo();
@@ -97,12 +97,11 @@ void sync(int device);
 bool synchronize_calls();
 
 int getActiveDeviceType();
+
 int getActivePlatform();
 
 bool& evalFlag();
 
-///////////////////////// BEGIN Sub-Managers ///////////////////
-//
 MemoryManager& memoryManager();
 
 MemoryManagerPinned& pinnedMemoryManager();
@@ -119,97 +118,23 @@ void addKernelToCache(int device, const std::string& key,
 void removeKernelFromCache(int device, const std::string& key);
 
 kc_entry_t kernelCache(int device, const std::string& key);
-//
-///////////////////////// END Sub-Managers /////////////////////
 
-class DeviceManager {
-    friend MemoryManager& memoryManager();
+static afcl::platform getPlatformEnum(cl::Device dev) {
+    std::string pname = getPlatformName(dev);
+    if (verify_present(pname, "AMD")) return AFCL_PLATFORM_AMD;
+    else if (verify_present(pname, "NVIDIA"))
+        return AFCL_PLATFORM_NVIDIA;
+    else if (verify_present(pname, "INTEL"))
+        return AFCL_PLATFORM_INTEL;
+    else if (verify_present(pname, "APPLE"))
+        return AFCL_PLATFORM_APPLE;
+    else if (verify_present(pname, "BEIGNET"))
+        return AFCL_PLATFORM_BEIGNET;
+    else if (verify_present(pname, "POCL"))
+        return AFCL_PLATFORM_POCL;
+    return AFCL_PLATFORM_UNKNOWN;
+}
 
-    friend MemoryManagerPinned& pinnedMemoryManager();
+void setActiveContext(int device);
 
-    friend graphics::ForgeManager& forgeManager();
-
-    friend GraphicsResourceManager& interopManager();
-
-    friend PlanCache& fftManager();
-
-    friend void addKernelToCache(int device, const std::string& key,
-                                 const kc_entry_t entry);
-
-    friend void removeKernelFromCache(int device, const std::string& key);
-
-    friend kc_entry_t kernelCache(int device, const std::string& key);
-
-    friend std::string getDeviceInfo();
-
-    friend int getDeviceCount();
-
-    friend int getDeviceIdFromNativeId(cl_device_id id);
-
-    friend const cl::Context& getContext();
-
-    friend cl::CommandQueue& getQueue();
-
-    friend const cl::Device& getDevice(int id);
-
-    friend size_t getDeviceMemorySize(int device);
-
-    friend bool isGLSharingSupported();
-
-    friend bool isDoubleSupported(int device);
-
-    friend void devprop(char* d_name, char* d_platform, char* d_toolkit,
-                        char* d_compute);
-
-    friend int setDevice(int device);
-
-    friend void addDeviceContext(cl_device_id dev, cl_context cxt,
-                                 cl_command_queue que);
-
-    friend void setDeviceContext(cl_device_id dev, cl_context cxt);
-
-    friend void removeDeviceContext(cl_device_id dev, cl_context ctx);
-
-    friend int getActiveDeviceType();
-
-    friend int getActivePlatform();
-
-   public:
-    static const unsigned MAX_DEVICES = 32;
-
-    static DeviceManager& getInstance();
-
-    ~DeviceManager();
-
-   protected:
-    DeviceManager();
-
-    // Following two declarations are required to
-    // avoid copying accidental copy/assignment
-    // of instance returned by getInstance to other
-    // variables
-    DeviceManager(DeviceManager const&);
-    void operator=(DeviceManager const&);
-    void markDeviceForInterop(const int device, const void* wHandle);
-
-   private:
-    // Attributes
-    common::mutex_t deviceMutex;
-    std::vector<cl::Device*> mDevices;
-    std::vector<cl::Context*> mContexts;
-    std::vector<cl::CommandQueue*> mQueues;
-    std::vector<bool> mIsGLSharingOn;
-    std::vector<int> mDeviceTypes;
-    std::vector<int> mPlatforms;
-    unsigned mUserDeviceOffset;
-
-    std::unique_ptr<graphics::ForgeManager> fgMngr;
-    std::unique_ptr<MemoryManager> memManager;
-    std::unique_ptr<MemoryManagerPinned> pinnedMemManager;
-    std::unique_ptr<GraphicsResourceManager> gfxManagers[MAX_DEVICES];
-    std::unique_ptr<clfftSetupData> mFFTSetup;
-
-    using BoostProgCache = boost::shared_ptr<boost::compute::program_cache>;
-    std::vector<BoostProgCache*> mBoostProgCacheVector;
-};
 }  // namespace opencl
